moved expr quad generation to new file
