+++ src/quads/quads.c
@@ -1,13 +1,12 @@
 #include <quads/quads.h>
 #include <quads/printutils.h>
 #include <quads/sizeof.h>
+#include <quads/exprquads.h>
 #include <string.h>
 #include <stdint.h>
 #include <parser.tab.h>
 
 // predeclaring some local functions
-static struct addr *gen_assign(union astnode *expr,
-	struct addr *target, struct basic_block *bb);
 static void generate_quads_rec(union astnode *stmt, struct basic_block *bb);
 
 // current function name and basic block number
@@ -15,12 +14,7 @@ static char *fn_name;
 static int bb_no, tmp_no;
 struct loop *cur_loop;
 
-/**
- * generates a new basic block with a unique identifier
- * 
- * @return		a new basic block
- */
-static struct basic_block *basic_block_new()
+struct basic_block *basic_block_new(void)
 {
 	struct basic_block *bb = calloc(1, sizeof(struct basic_block));
 
@@ -31,19 +25,7 @@ static struct basic_block *basic_block_new()
 	return bb;
 }
 
-/**
- * emits a new quad to the specified basic block
- * 
- * any of the operands or src may be null, depending on the opcode
- * 
- * @param bb		basic block to emit quad to
- * @param opcode	quad opcode
- * @param dest		quad destination; must be an lvalue
- * @param src1		quad first operand
- * @param src2		quad second operand
- * @return		generated quad
- */
-static struct quad *quad_new(struct basic_block *bb, enum opcode opcode,
+struct quad *quad_new(struct basic_block *bb, enum opcode opcode,
 	struct addr *dest, struct addr *src1, struct addr *src2)
 {
 	struct quad *quad = calloc(1, sizeof(struct quad));
@@ -64,15 +46,7 @@ static struct quad *quad_new(struct basic_block *bb, enum opcode opcode,
 	return quad;
 }
 
-/**
- * constructs and returns a new struct addr (operand/dest to quad)
- * 
- * @param type		type of addr (memory (variable), immediate (constant),
- * 			or temporary (register))
- * @param decl		astnode representation of the type of the value
- * @return		constructed struct addr
- */
-static struct addr *addr_new(enum addr_type type, union astnode *decl)
+struct addr *addr_new(enum addr_type type, union astnode *decl)
 {
 	struct addr *addr = calloc(1, sizeof(struct addr));
 
@@ -85,15 +59,7 @@ static struct addr *addr_new(enum addr_type type, union astnode *decl)
 	return addr;
 }
 
-/**
- * constructs and returns a new temporary pseudo-register (for subexpressions)
- * 
- * also gives the struct addr a unique ID
- * 
- * @param decl		astnode representation of the type of the value
- * @return		constructed struct addr
- */
-static struct addr *tmp_addr_new(union astnode *decl)
+struct addr *tmp_addr_new(union astnode *decl)
 {
 	struct addr *addr = addr_new(AT_TMP, decl);
 
@@ -102,427 +68,6 @@ static struct addr *tmp_addr_new(union astnode *decl)
 
 	return addr;
 }
-
-/**
- * iteratively and recursively generates a linked-list of quads for an
- * (r-value) expression; corresponds to function of same name in lecture notes
- * 
- * The use of the target address is to have more optimized 3-address quads,
- * i.e., remove extra MOV quads. (These will still have to be added back in
- * when we get to assembly code generation, as x86 is a 2-address architecture.)
- * 
- * semantic notes:
- * - currently no support for floating point in expressions, only simple
- * 	scalar types (char, ([unspec]|short|long|long long) int)
- * - currently no support for signed/unsigned type conversions
- * - if primary expression (scalar or constant):
- * 	- if no target, return as a struct addr
- * 	- if target is specified, emit MOV opcode
- * - if assignment (=): see gen_assign()
- * - if unary/binary expression, emit quad
- * 	- if dest is NULL, a temporary is created
- *	- if target, it is the destination of the quad
- *
- * @param expr		expression object
- * @param dest		target address, or NULL if temporary expression
- * @param bb		basic block to add quads to
- * @return 		addr storing result of expression
- */
-static struct addr *gen_rvalue(union astnode *expr,
-	struct addr *dest, struct basic_block *bb)
-{
-	struct addr *src1, *src2, *tmp, *tmp2;
-	union astnode *ts;
-	enum opcode op;
-	unsigned subtype_size;
-
-	// null expression
-	// this shouldn't happen but this is here as a safety measure
-	if (!expr) {
-		yyerror("quadgen: empty expression in gen_rvalue()");
-		return NULL;
-	}
-
-	switch (NT(expr)) {
-
-	// symbol
-	case NT_DECL:
-		// if abstract, don't do anything and don't return anything
-		if (!expr->decl.ident) {
-			return NULL;
-		}
-
-		// don't support expressions with non-integral types
-		if (NT(expr->decl.components) == NT_DECLSPEC) {
-			ts = expr->decl.components->declspec.ts;
-			
-			if (NT(ts) == NT_TS_SCALAR
-				&& ts->ts_scalar.basetype != BT_INT
-				&& ts->ts_scalar.basetype != BT_CHAR) {
-				yyerror_fatal("only int, char types allowed in"
-					" expressions at this time");
-			}
-		}
-		
-		// treat array as pointer (special cases are treated elsewhere)	
-		if (NT(expr->decl.components) == NT_DECLARATOR_ARRAY) {
-			// TODO: maybe (?) change type to be that of a pointer
-			// TODO: in add/subtract/assign array, downgrade it
-			// 	and write a function to do this
-
-			// ALLOC_TYPE(ts, NT_DECLARATOR_POINTER);
-			// ts->decl_pointer.of =
-			// 	expr->decl.components->decl_array.of;
-			// src1 = addr_new(AT_AST, ts);
-			src1 = addr_new(AT_AST, expr->decl.components);
-			src1->val.astnode = expr;
-
-			if (!dest) {
-				dest = tmp_addr_new(src1->decl);
-			}
-			quad_new(bb, OC_LEA, dest, src1, NULL);
-		}
-		
-		// not a pointer
-		else {
-			src1 = addr_new(AT_AST, expr->decl.components);
-			src1->val.astnode = expr;
-
-			if (dest) {
-				quad_new(bb, OC_MOV, dest, src1, NULL);
-			} else {
-				dest = src1;
-			}
-		}
-
-		return dest;
-
-	// constant number
-	case NT_NUMBER:
-		// don't support expressions with non-integral types
-		if (expr->num.ts->ts_scalar.basetype != BT_INT
-			&& expr->num.ts->ts_scalar.basetype != BT_CHAR) {
-			yyerror_fatal("only int, char types allowed in"
-					" expressions at this time");
-		}
-
-		// convert into const
-		src1 = addr_new(AT_CONST, expr->num.ts);
-		*((uint64_t *)src1->val.constval) =
-			*((uint64_t *)expr->num.buf);
-
-		if (dest) {
-			quad_new(bb, OC_MOV, dest, src1, NULL);
-		} else {
-			dest = src1;
-		}
-		return dest;
-
-	// unary operator
-	case NT_UNOP:
-		src1 = gen_rvalue(expr->unop.arg, NULL, bb);
-
-		// TODO: still have to implement a lot here
-		switch (expr->unop.op) {
-
-		// sizeof with a symbol or constexpr
-		// TODO: note that sizeof returns type size_t, defined in
-		// 	stddef.h -- should represent this type somewhere
-		case SIZEOF:
-			// create 8-byte type like size_t (similar to
-			// unsigned long long in 64-bit systems)
-			ALLOC_TYPE(ts, NT_TS_SCALAR);
-			ts->ts_scalar.basetype = BT_INT;
-			ts->ts_scalar.modifiers.lls = LLS_LONG_LONG;
-			ts->ts_scalar.modifiers.sign = SIGN_UNSIGNED;
-
-			src2 = addr_new(AT_CONST, ts);
-			*((uint64_t *)src2->val.constval)
-				= src1->type == AT_CONST
-				? src1->size
-				: astnode_sizeof_type(src1->decl);
-			
-			if (dest) {
-				quad_new(bb, OC_MOV, dest, src2, NULL);
-			} else {
-				dest = src2;
-			}
-			return dest;
-
-		// sizeof with a typename (addr1 should be NULL)
-		case 's':
-			// see notes above
-			ALLOC_TYPE(ts, NT_TS_SCALAR);
-			ts->ts_scalar.basetype = BT_INT;
-			ts->ts_scalar.modifiers.lls = LLS_LONG_LONG;
-			ts->ts_scalar.modifiers.sign = SIGN_UNSIGNED;
-
-			src1 = addr_new(AT_CONST, ts);
-			*((uint64_t *)src1->val.constval)
-				= astnode_sizeof_type(expr->unop.arg
-					->decl.components);
-
-			if (dest) {
-				quad_new(bb, OC_MOV, dest, src1, NULL);
-			} else {
-				dest = src1;
-			}
-			return dest;
-
-		// pointer deref
-		case '*':
-			// check that the rvalue is a pointer type
-			if (NT(src1->decl) != NT_DECLARATOR_POINTER
-				&& NT(src1->decl) != NT_DECLARATOR_ARRAY) {
-				yyerror_fatal("dereferencing non-pointer type");
-			}
-
-			// create new addr of underlying type to store the
-			// result in
-			if (!dest) {
-				// get the type that the pointer is pointing to
-				ts = src1->decl->decl_pointer.of;
-				dest = tmp_addr_new(ts);
-
-				// noop "pseudo-mov," i.e., reinterpret pointer
-				// as different size but same pointer value
-				quad_new(bb, OC_PMOV, dest, src1, NULL);
-			}
-
-			// noop if pointer to/array of array
-			if (NT(src1->decl->decl_array.of)
-				!= NT_DECLARATOR_ARRAY) {
-				quad_new(bb, OC_LOAD, dest, src1, NULL);
-			}
-			return dest;
-		}
-		break;
-
-	// binary operator
-	case NT_BINOP:
-		switch(expr->binop.op){
-		case '+':	op = OC_ADD;	break;
-		case '-':	op = OC_SUB;	break;
-		case '*':	op = OC_MUL;	break;
-		case '/':	op = OC_DIV;	break;
-		case '%':	op = OC_MOD;	break;
-		case '&':	op = OC_AND;	break;
-		case '|':	op = OC_OR;	break;
-		case '^':	op = OC_XOR;	break;
-		case SHL:	op = OC_SHL;	break;
-		case SHR:	op = OC_SHR;	break;
-
-		// special binop: assignment
-		case '=':	return gen_assign(expr, dest, bb);
-		}
-
-		src1 = gen_rvalue(expr->binop.left, NULL, bb);
-		src2 = gen_rvalue(expr->binop.right, NULL, bb);
-
-		// TODO: +/- have to correctly implement pointer arithmetic
-		// 	(have to check type of their operands, which means
-		// 	that we have to associate type with struct addrs)
-		switch (expr->binop.op) {
-		// arithmetic
-		case '+':
-			// helper: is array or pointer
-			#define AOP(addr) (NT(addr->decl) == NT_DECLARATOR_ARRAY \
-				|| NT(addr->decl) == NT_DECLARATOR_POINTER)
-
-			// make the first one the pointer, if applicable
-			if (AOP(src2)) {
-				tmp = src2;
-				src2 = src1;
-				src1 = tmp;
-			}
-
-			// error case: pointer + pointer
-			if (AOP(src1) && AOP(src2)) {
-				yyerror_fatal("pointer + pointer");
-				return NULL;
-			}
-
-			// special case: pointer + int;
-			if (AOP(src1)) {
-				// insert operation to multiply src2 by sizeof
-				// src1 subtype
-
-				// see notes above
-				ALLOC_TYPE(ts, NT_TS_SCALAR);
-				ts->ts_scalar.basetype = BT_INT;
-				ts->ts_scalar.modifiers.lls = LLS_LONG_LONG;
-				ts->ts_scalar.modifiers.sign = SIGN_UNSIGNED;
-
-				// convert into const
-				tmp = addr_new(AT_CONST, ts);
-				*((uint64_t *)tmp->val.constval) =
-					astnode_sizeof_type(src1->decl->
-						decl_pointer.of);
-
-				tmp2 = tmp_addr_new(src2->decl);
-				quad_new(bb, OC_MUL, tmp2, tmp, src2);
-				src2 = tmp2;
-			}
-
-			// create new tmp
-			// TODO: choose larger of two sizes
-			// 	(or better yet, use real types rather than just
-			// 	sizes)
-			if (!dest) {
-				dest = tmp_addr_new(src1->decl);
-			}
-			quad_new(bb, OC_ADD, dest, src1, src2);
-			return dest;
-			
-		case '-':
-			if (!dest) {
-				dest = tmp_addr_new(src1->decl);
-			}
-			quad_new(bb, OC_SUB, dest, src1, src2);
-			return dest;
-
-		case '*':
-			if (!dest) {
-				dest = tmp_addr_new(src1->decl);
-			}
-			quad_new(bb, OC_MUL, dest, src1, src2);
-			return dest;
-		
-		case '/':
-			if (!dest) {
-				dest = tmp_addr_new(src1->decl);
-			}
-			quad_new(bb, OC_MUL, dest, src1, src2);
-			return dest;
-		
-		}
-		break;
-
-	default:
-		NYI("quadgen: other expression type quad generation");
-	}
-
-	yyerror_fatal("quadgen: invalid fallthrough; some expr not handled");
-	return NULL;
-}
-
-/**
- * generate a lvalue (in the case of variable assignment)
- * 
- * semantic notes:
- * - only is successful if expr is a variable (memory address) or deref
- * - currently only scalar variables are supported, struct/union assignments
- *  	are not
- * 
- * @param expr		expression astnode representing lvalue
- * @param bb		containing basic block
- * @param mode		set addressing mode
- * @return		generated lvalue
- */
-static struct addr *gen_lvalue(union astnode *expr,
-	struct basic_block *bb, enum addr_mode *mode)
-{
-	struct addr *dest;
-
-	switch (NT(expr)) {
-	
-	// variable
-	case NT_DECL:
-
-		// if abstract, not an lvalue; fallthrough to error
-		if (!expr->decl.ident) {
-			break;
-		}
-
-		switch (NT(expr->decl.components)) {
-
-		case NT_DECLSPEC:
-			if (NT(expr->decl.components->declspec.ts) !=
-				NT_TS_SCALAR) {
-				yyerror_fatal("assignment to struct/union"
-					" not supported (yet)");
-				return NULL;
-			}
-		case NT_DECLARATOR_POINTER:
-
-			*mode = AM_DIRECT;
-
-			dest = addr_new(AT_AST, expr->decl.components);
-			dest->val.astnode = expr;
-			return dest;
-		}
-
-		// fallthrough: invalid lvalue (shouldn't happen)
-		break;
-
-	// deref
-	case NT_UNOP:
-		// only valid operation that returns an lvalue is deref,
-		// if not fallthrough to error
-		if (expr->unop.op != '*') {
-			break;
-		}
-
-		*mode = AM_INDIRECT;
-
-		dest = gen_rvalue(expr->unop.arg, NULL, bb);
-
-		// check that the rvalue is a pointer type
-		if (NT(dest->decl) != NT_DECLARATOR_POINTER
-			&& NT(dest->decl) != NT_DECLARATOR_ARRAY) {
-			yyerror_fatal("dereferencing non-pointer type");
-		}
-
-		return dest;
-	}
-
-	yyerror_fatal("quadgen: gen_lvalue fallthrough: invalid lvalue");
-	return NULL;
-}
-
-/**
- * handle assignment expressions
- * 
- * semantic notes:
- * - returns value in case it is used as an rvalue (note: assignment is an
- * 	rvalue in C (section 6.5.16) but an lvalue in C++ (section 5.17))
- * - if multiple chained assignments (i.e., dest is specified), then will also
- * 	emit a MOV instruction
- * 
- * @param expr		assignment astnode expression
- * @param target	destination if directly chained assignment to a variable
- * @param bb		containing basic block
- * @return		addr of RHS of expression (for use as an rvalue)
- */
-static struct addr *gen_assign(union astnode *expr,
-	struct addr *target, struct basic_block *bb)
-{
-	struct addr *dest, *src;
-	enum addr_mode mode;
-
-	if (NT(expr) != NT_BINOP || expr->binop.op != '=') {
-		yyerror_fatal("quadgen: gen_assign(): not an assignment");
-		return NULL;
-	}
-
-	// generate lvalue; if invalid lvalue, will report and panic in
-	// gen_lvalue()
-	dest = gen_lvalue(expr->binop.left, bb, &mode);
-	if (mode == AM_DIRECT) {
-		src = gen_rvalue(expr->binop.right, dest, bb);
-	} else {
-		src = gen_rvalue(expr->binop.right, NULL, bb);
-		quad_new(bb, OC_STORE, NULL, src, dest);
-	}
-
-	// directly being assigned to a memory location, issue a MOV
-	if (target) {
-		quad_new(bb, OC_MOV, NULL, src, target);
-	}
-
-	return src;
-}
-
 
 /*Loops*/
